requirements.txt
streamlit==1.34.0
pandas>=2.0.0
gspread>=5.10.0
google-auth>=2.20.0
google-auth-oauthlib>=1.2.0
plotly>=5.19.0
python-dateutil>=2.8.2
numpy>=1.23.0
requests>=2.28.0